requirements.txt
MarkupSafe==1.1.1
Click==7.0
Flask==1.0.3
future==0.17.1
google-pasta==0.1.7
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.16.2
Pillow==6.0.0
requests==2.20.1
tensorboard==1.13.1
tensorflow==1.14.0rc1
tensorflow-estimator==1.14.0rc1
Werkzeug==0.15.4
gunicorn==19.9.0